Moved fatal error test support into a separate library.
